Migration 37
1. Исправлен баг с неправильным формированием запроса функции 0x10 Modbus.
2. Рефакторинг.
3. Все VM режимов записи теперь формируют данные.

+++ ViewModels/ModbusClient/DataTypes/ModbusDataDisplayed.cs
@@ -6,7 +6,7 @@
         public string? FuncNumber { get; set; }
         public ushort Address { get; set; }
         public string? ViewAddress { get; set; }
-        public ushort[]? Data { get; set; }
+        public byte[]? Data { get; set; }
         public string? ViewData { get; set; }
     }
 }
